Rename his directory to analysis
