+++ src/components/analysis/DataAnalysisDisplay.vue
