+++ src/components/analysis/HisCollectionChart.vue
