+++ src/components/analysis/HisCollectionCharts.vue
