+++ src/components/analysis/HisCorrelationChart.vue
