+++ src/router/index.ts
@@ -39,7 +39,7 @@ const WebDisplay = () => import('../components/webdisplay/WebDisplay.vue')
 const DashboardView = () => import('../views/DashboardView.vue')
 const LogDisplay = () => import('../components/logdisplay/LogDisplay.vue')
 const DataAnalysisDisplay = () =>
-  import('../components/his/DataAnalysisDisplay.vue')
+  import('../components/analysis/DataAnalysisDisplay.vue')
 const Empty = () => import('../views/Empty.vue')
 const DocumentDisplayView = () => import('../views/DocumentDisplayView.vue')
